+++ src/RxCollection.js
@@ -6,25 +6,24 @@ import * as util from './util';
 import RxDocument from './RxDocument';
 import RxQuery from './RxQuery';
 import RxChangeEvent from './RxChangeEvent';
-import KeyCompressor from './KeyCompressor';
 import DataMigrator from './DataMigrator';
 import Crypter from './Crypter';
 import DocCache from './DocCache';
 import QueryCache from './QueryCache';
 import ChangeEventBuffer from './ChangeEventBuffer';
 import RxReplicationState from './RxReplicationState';
+import overwritable from './overwritable';
+import {
+    runPluginHooks
+} from './hooks';
 
-import {
-    RxSchema
-} from './RxSchema';
-import {
-    RxDatabase
-} from './RxDatabase';
+import RxSchema from './RxSchema';
+import RxDatabase from './RxDatabase';
 
 const HOOKS_WHEN = ['pre', 'post'];
 const HOOKS_KEYS = ['insert', 'save', 'remove', 'create'];
 
-class RxCollection {
+export class RxCollection {
     constructor(database, name, schema, pouchSettings = {}, migrationStrategies = {}, methods = {}) {
         this.database = database;
         this.name = name;
@@ -56,7 +55,6 @@ class RxCollection {
     async prepare() {
         this._dataMigrator = DataMigrator.create(this, this._migrationStrategies);
         this._crypter = Crypter.create(this.database.password, this.schema);
-        this._keyCompressor = KeyCompressor.create(this.schema);
 
         this.pouch = this.database._spawnPouchDB(this.name, this.schema.version, this._pouchSettings);
         this._observable$ = this.database.$
@@ -69,9 +67,10 @@ class RxCollection {
             .map(indexAr => {
                 const compressedIdx = indexAr
                     .map(key => {
-                        if (!this.schema.doKeyCompression()) return key;
-                        const ret = this._keyCompressor._transformKey('', '', key.split('.'));
-                        return ret;
+                        if (!this.schema.doKeyCompression())
+                            return key;
+                        else
+                            return this._keyCompressor._transformKey('', '', key.split('.'));
                     });
                 return this.pouch.createIndex({
                     index: {
@@ -81,7 +80,6 @@ class RxCollection {
             })
         );
 
-
         this._subs.push(
             this._observable$.subscribe(cE => {
                 // when data changes, send it to RxDocument in docCache
@@ -89,6 +87,12 @@ class RxCollection {
                 if (doc) doc._handleChangeEvent(cE);
             })
         );
+    }
+
+    get _keyCompressor() {
+        if (!this.__keyCompressor)
+            this.__keyCompressor = overwritable.createKeyCompressor(this.schema);
+        return this.__keyCompressor;
     }
 
     /**
@@ -122,17 +126,21 @@ class RxCollection {
      * wrappers for Pouch.put/get to handle keycompression etc
      */
     _handleToPouch(docData) {
-        const encrypted = this._crypter.encrypt(docData);
-        const swapped = this.schema.swapPrimaryToId(encrypted);
-        const compressed = this._keyCompressor.compress(swapped);
-        return compressed;
+        let data = clone(docData);
+        data = this._crypter.encrypt(data);
+        data = this.schema.swapPrimaryToId(data);
+        if (this.schema.doKeyCompression())
+            data = this._keyCompressor.compress(data);
+        return data;
     }
     _handleFromPouch(docData, noDecrypt = false) {
-        const swapped = this.schema.swapIdToPrimary(docData);
-        const decompressed = this._keyCompressor.decompress(swapped);
-        if (noDecrypt) return decompressed;
-        const decrypted = this._crypter.decrypt(decompressed);
-        return decrypted;
+        let data = clone(docData);
+        data = this.schema.swapIdToPrimary(data);
+        if (this.schema.doKeyCompression())
+            data = this._keyCompressor.decompress(data);
+        if (noDecrypt) return data;
+        data = this._crypter.decrypt(data);
+        return data;
     }
 
     /**
@@ -709,10 +717,10 @@ export async function create({
     statics = {},
     methods = {}
 }) {
-    if (!schema instanceof RxSchema)
+    if (!RxSchema.isInstanceOf(schema))
         throw new TypeError('given schema is no Schema-object');
 
-    if (!database instanceof RxDatabase)
+    if (!RxDatabase.isInstanceOf(database))
         throw new TypeError('given database is no Database-object');
 
     if (typeof autoMigrate !== 'boolean')
@@ -743,6 +751,7 @@ export async function create({
     if (autoMigrate)
         await collection.migratePromise();
 
+    runPluginHooks('createRxCollection', collection);
     return collection;
 }
 
@@ -753,5 +762,6 @@ export function isInstanceOf(obj) {
 export default {
     create,
     properties,
-    isInstanceOf
+    isInstanceOf,
+    RxCollection
 };