+++ src/RxDatabase.js
@@ -6,15 +6,12 @@ import RxCollection from './RxCollection';
 import RxSchema from './RxSchema';
 import RxChangeEvent from './RxChangeEvent';
 import Socket from './Socket';
-import LeaderElector from './LeaderElector';
-
-const SETTINGS = {
-    minPassLength: 8
-};
+import overwritable from './overwritable';
+import {
+    runPluginHooks
+} from './hooks';
 
 export class RxDatabase {
-
-
     constructor(name, adapter, password, multiInstance) {
         this.name = name;
         this.adapter = adapter;
@@ -75,9 +72,12 @@ export class RxDatabase {
             //TODO only subscribe when sth is listening to the event-chain
             this.socket.messages$.subscribe(cE => this.$emit(cE));
         }
+    }
 
-        // leader elector
-        this.leaderElector = await LeaderElector.create(this);
+    get leaderElector() {
+        if (!this._leaderElector)
+            this._leaderElector = overwritable.createLeaderElector(this);
+        return this._leaderElector;
     }
 
     /**
@@ -369,14 +369,13 @@ export async function create({
         }
     }
 
-    if (password && typeof password !== 'string')
-        throw new TypeError('password is no string');
-    if (password && password.length < SETTINGS.minPassLength)
-        throw new Error(`password must have at least ${SETTINGS.minPassLength} chars`);
+    if (password)
+        overwritable.validatePassword(password);
 
     const db = new RxDatabase(name, adapter, password, multiInstance);
     await db.prepare();
 
+    runPluginHooks('createRxDatabase', db);
     return db;
 }
 
@@ -456,14 +455,15 @@ export function isInstanceOf(obj) {
     return obj instanceof RxDatabase;
 }
 
+// TODO is this needed?
 export {
     RxSchema as RxSchema
 };
-
 
 export default {
     create,
     removeDatabase,
     isInstanceOf,
+    RxDatabase,
     RxSchema
 };